Limit the maximum number of retries for failed actions
Summary:
Currently a failed action will run until it works -- for some of them, it means running forever. This diff limits the number of retries to 20 and adds a specific log message when giving up on an action.

It should be interesting to have add to the kibana dashboard a panel with the top failing actions and accounts.

Test Plan:
Ran an always failing action -- checked that it worked
Ran a working action -- checked that it worked

Reviewers: kav-ya, emfree

Reviewed By: kav-ya, emfree

Projects: #sync-engine

Differential Revision: https://review.inboxapp.com/D562