Fix uidvalidity_valid check.
Summary:
It appears that server UIDVALIDITY values sometimes *decrease*.  Per RFC3501
the server is supposed to *increase* the UIDVALIDITY if UIDS are actually
invalid. So it should in theory be safe to ignore a decremented UIDVALIDITY? In
practice, that appears to be the case for the actually affected accounts.

Anyways, flip the comparison. Fixes T471, sort of.

Test Plan: Stare at code; RFC.

Reviewers: spang

Reviewed By: spang

Maniphest Tasks: T471

Differential Revision: https://review.inboxapp.com/D583