Support account (and namespace) deletion.
Summary:
This revision introduces account (and namespace) deletion support. It makes a
few changes to our models (mostly on cascading operations) either at the DBMS-
or the ORM-level.

NOTE: this revision also applies migration version 127 (it was missing).

Test Plan: Test the migration on a snapshot of the prod db to get insights about timing.

Reviewers: emfree

Subscribers: spang

Differential Revision: https://review.inboxapp.com/D1025