[WIP] Contacts refactor.
Summary:
Note this is an incomplete revision. Specifically,
- events sync needs to be updated (the current changes to base_poll break it).
- Tests need to be updated.
- Migrations needed.

The reason then for this revision is merely to ask if you'd envisioned
doing the refactor a different way.

I believe the code can be further simplified but I'll have a more specific idea
once the events refactor is underway.

Test Plan: None, WIP revision.

Reviewers: emfree

Differential Revision: https://review.inboxapp.com/D1036