Better handling of deleted folders within FolderSyncEngine
Summary:
There are some cases where the folder a FolderSyncEngine is operating on could be deleted while the folder is running. This diff handles those cases from within the sync engine, raising MailsyncDone appropriately.

The first case is that the folder row gets deleted from the database while the sync is starting up, causing an IntegrityError when the sync engine tries to insert its sync status row (T881). This has been observed live repeatedly with Yahoo accounts auto-deleting the Y! Conversations folder.

The second case is that the crispin connection fails on `select_folder` with a specific 'Folder does not exist' message. This can happen any time a `select_folder` is performed in the main sync loop, and has again been observed live with Y! Conversations folders.

We handle these cases by watching for `IntegrityError` in the former case and having crispin throw a `FolderMissingError` in the latter. Both raise `MailsyncDone` and we let the monitor logic deal with deleting the folder rows if necessary, and recreating sync engines as required.

Test Plan: Added test coverage to make sure the new errors behave as expected. Am attempting to manually trigger the Y! specific bug on local dev, but would probably benefit from more extended run testing.

Reviewers: kav-ya

Reviewed By: kav-ya

Subscribers: khamidou, spang, emfree

Differential Revision: https://review.inboxapp.com/D1442