Deal with user created folders with reserved (canonical) names
Summary:
Users may have folders with the same names as canonical folders. This can cause IntegrityErrors and false tag deletions (T1079), since we key off name alone.

This diff updates `Folder.get_associated_tag` to check the `public_id` on the tag as well as the name if it's not a canonical tag, thereby ensuring we don't return the canonical tag when we really should be creating a user-local tag. As a result, the unique constraint on Tag has to change too.

Test Plan: Added a unit test to cover this (and a second to check for a race condition that I suspected was happening but appears not to be -- but seems worth keeping around). Also triggered this issue manually by adding a folder 'spam' on a Gmail account and then deleting it.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://review.inboxapp.com/D1469