Fix T1113: Data too long for column owner
Summary:
This is another bug caused by differences between our prod db and the schema our code expects. This time, it’s because the `owner` column defined on prod is a `varchar(255)` while our code expects it to be a `varchar(1024)`. This diff adds a migration to fix this.

Please note that the diff is based off master because I noticed our db dump had the same problem. Given that `ALTER TABLE` seems to be idempotent with MySQL (i.e when running the statement twice, the second time MySQL doesn’t match any column), I thought it would be okay.

Test Plan:
Tested manually:
- ran the migration inside mysql
- upgraded a db dump

Reviewers: spang, emfree, jennie

Reviewed By: spang

Projects: #sync-engine

Maniphest Tasks: T1113

Differential Revision: https://phab.nylas.com/D1488