Fix the race-condition preventing updating an EAS event more than once
Summary:
This bug was tricky to debug. So, according to the report, updating an event /once/ would work but not twice, which is super weird.

My first step was looking at the EAS action log and stepping through the code when syncing back an event.
After a lot of head scratching I noticed that we weren’t even going through this portion of the code. That’s when it dawned on me that we had a little piece of code that would « collapse » updates to the same object. Lo and behold, the code in question isn’t aware of the fact that EAS events have both a `status` and `secondary_status` column, and would helpfully collapse events for us. Because more often than not, the secondary device has run before the first, we would have an event where `status=pending` and `secondary_status=successful` which made our EAS action log handler ignore the update.

Also, this is unrelated but why are we syncing back anything for the primary device?

Test Plan: Tested manually --- checked that updating an event twice worked.

Reviewers: kav-ya, emfree

Subscribers: spang

Projects: #eas, #calendars

Maniphest Tasks: T1826

Differential Revision: https://phab.nylas.com/D1593