Prevent dropped UIDs in IMAP sync.
Summary:
Previously, Gmail All Mail UIDs would not be saved if they referenced a
message which already existed (as identified by its g_msgid). This is an
artefact of the old Inbox/All Mail split. But it meant that, for example, UIDs
would not properly be committed for messages moved from the Trash to the All
Mail folder. Fix this.

Also fix race condition in managing pending UIDs. Previously, code that
downloaded uids looked like this:
```
while not download_stack.empty():
    uid, metadata = download_stack.peek()
    self.download_and_commit_uids(crispin_client, self.folder_name,
                                  [uid])
    download_stack.get()
```

The idea was that if you just popped the UID off the stack, it could end up
being concurrently added back to the stack by the `poll_for_changes` greenlet
while the download is ongoing.

However, it can happen that a different UID is concurrently added to the stack.
So `download_stack.get()` would remove the wrong UID. Fix this too.

Test Plan:
Introduce latency and debug logging in the original code; observe
the condition described.

Reviewers: spang, kav-ya

Reviewed By: kav-ya

Differential Revision: https://phab.nylas.com/D1667