Add column initial_synced_at to Folder that reports when initial_sync is done
Summary:
This must be done at the folder level because the initial/poll/etc state occurs
at the folder level. To help determine when an account has completed initial
sync, I've added a @property function that looks at the folders for that
account to determine when all the folders completed that sync

Test Plan: Looking at the current tests, it was not obvious to me how to add tests in the situation where syncs are actually run. We have tests that create foldersyncengines but I couldn't find any that ran the syncs. Given this is only for metrics, the code is simple, and it doesn't break current tests, I thought this is good enough. If you disagree I can spend more time looking into adding a test.

Reviewers: emfree

Reviewed By: emfree

Differential Revision: https://phab.nylas.com/D1733