Add minimal detection of excessive lazy-loading.
Summary:
Basically just keep track of how many queries are issued with a given
connection before commit() is called, and log a warning if that number is
high. This is a good way to tell if, for example, lots of things are being
lazily loaded. Also bind more logging context in the API, so that we can tell
which endpoints are doing these bad things.

Fixes T1098.

Test Plan: Run API. Issue some requests. Inspect logs. Recoil in horror.

Reviewers: spang

Reviewed By: spang

Maniphest Tasks: T1098

Differential Revision: https://phab.nylas.com/D1788