Use database name from config in pool instrumentation.
Summary:
If we opt to use MySQLdb rather than PyMySQL (as we might for result-fetching
performance), then `dbapi_connection` in the event listener is an instance of
`MySQLdb.connections.Connection` and does not expose a `db` attribute. So
instead just grab the database name from the config.

Test Plan:
change database URI to `mysql+mysqldb://...` and check that it
works.

Reviewers: systemizer

Reviewed By: systemizer

Differential Revision: https://phab.nylas.com/D1845